function for packing data message implemented by client, not api class
